fix(crc): avoid integer underflow for empty CRC data
Empty data being fed to the `extract_crc` function causes an integer
underflow and resulting `panic` when used in downstream projects.

While the data being empty
int the first place is a somewhat pathological issue and suggests a
problem with external tooling/downstream users, we can easily avoid
`panic`s in the case of that pathological case.

This commit updates the CRC parsing code to account for the
possibility that provided data is empty.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>